requirements.txt
discord.py
docker
requests